Fix enum names that start with an underscore.
