+++ pkg/codegen/dotnet/utilities.go
@@ -102,6 +102,11 @@ func makeSafeEnumName(name string) (string, error) {
 	// Capitalize and make a valid identifier.
 	safeName = strings.Title(makeValidIdentifier(safeName))
 
+	// If the first character of the name is an underscore, replace it with `Value`.
+	if strings.HasPrefix(safeName, "_") {
+		safeName = strings.Replace(safeName, "_", "Value", 1)
+	}
+
 	// If there are multiple underscores in a row, replace with one.
 	regex := regexp.MustCompile(`_+`)
 	safeName = regex.ReplaceAllString(safeName, "_")
