+++ pkg/codegen/dotnet/utilities_test.go
@@ -11,7 +11,8 @@ func TestMakeSafeEnumName(t *testing.T) {
 		{"+", "", true},
 		{"*", "Asterisk", false},
 		{"0", "Zero", false},
-		{"8.3", "_8_3", false},
+		{"8.3", "Value8_3", false},
+		{"11", "Value11", false},
 		{"Microsoft-Windows-Shell-Startup", "Microsoft_Windows_Shell_Startup", false},
 		{"Microsoft.Batch", "Microsoft_Batch", false},
 		{"readonly", "@Readonly", false},
