+++ pkg/codegen/python/gen_test.go
@@ -33,42 +33,6 @@ func TestRelPathToRelImport(t *testing.T) {
 	}
 }
 
-func TestMakeSafeEnumName(t *testing.T) {
-	tests := []struct {
-		input    string
-		expected string
-		wantErr  bool
-	}{
-		{"red", "RED", false},
-		{"snake_cased_name", "SNAKE_CASED_NAME", false},
-		{"+", "", true},
-		{"*", "ASTERISK", false},
-		{"0", "ZERO", false},
-		{"8.3", "_8_3", false},
-		{"Microsoft-Windows-Shell-Startup", "MICROSOFT_WINDOWS_SHELL_STARTUP", false},
-		{"Microsoft.Batch", "MICROSOFT_BATCH", false},
-		{"readonly", "READONLY", false},
-		{"SystemAssigned, UserAssigned", "SYSTEM_ASSIGNED_USER_ASSIGNED", false},
-		{"Dev(NoSLA)_Standard_D11_v2", "DEV_NO_SL_A_STANDARD_D11_V2", false},
-		{"Standard_E8as_v4+1TB_PS", "STANDARD_E8AS_V4_1_T_B_PS", false},
-		{"Plants'R'Us", "PLANTS_R_US", false},
-		{"Pulumi Planters Inc.", "PULUMI_PLANTERS_INC_", false},
-		{"ZeroPointOne", "ZERO_POINT_ONE", false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.input, func(t *testing.T) {
-			got, err := makeSafeEnumName(tt.input)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("makeSafeEnumName() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.expected {
-				t.Errorf("makeSafeEnumName() got = %v, want %v", got, tt.expected)
-			}
-		})
-	}
-}
-
 func TestGeneratePackage(t *testing.T) {
 	tests := []struct {
 		name          string
